requirements.txt
django
psycopg2
djangorestframework
djangorestframework-simplejwt
django-cors-headers
django-filter
pillow
markdown
django-import-export
future
httplib2
requests
coreapi

